move model selector to models panel
